+++ lib/high_level.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "faraday"
 require "zeitwerk"
 
 module HighLevel
@@ -8,4 +9,7 @@ end
 loader = Zeitwerk::Loader.new
 loader.tag = "gohighlevel"
 loader.push_dir("#{__dir__}/high_level", namespace: HighLevel)
+loader.ignore("#{__dir__}/high_level/errors.rb")
 loader.setup
+
+require_relative "high_level/errors"